backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
httpx>=0.27.0
schedule>=1.2.0
pyyaml>=6.0.0
